fix(app): a negotiated discount survives not being a whole percentage
Found by driving the flow on a device, not by reading it.

NegotiatedDeal.percent was an int parsed with int.tryParse, and
int.tryParse("9.38") is null — so it fell to 0. A negotiated price almost never
lands on a whole percentage: agreeing ₹2,900 on a ₹3,200 listing is 9.375%,
which the server returns as 9.38. Every such deal therefore reached the app as
"0% off". The banner said 0% OFF, and "Book at the agreed price" opened checkout
at the FULL price with the coupon attached and nothing taken off it — the guest
having just been told the price was agreed. Only a deal that happened to be a
round number ever worked.

percent is a double now, in NegotiatedDeal, in CouponValidation (the same
int.tryParse on /user/coupons/validate) and through openPropertyById into the
property page. Percentages render as a person would write them — "9.38", "10",
never "10.0".

Verified on the device against a real negotiation: offered ₹2,400 on Azure Sky
Apartment, host countered ₹2,900, accepted in-app, which minted DEAL9101C39 at
9.38%. Checkout now reads ₹6,400 base, −₹600 discount, ₹290 GST, ₹6,090 total —
which is the agreed ₹2,900 × 2 nights plus tax, to the rupee.

Two more, both found the same way:

_bookDeal raised a GetX snackbar and then called openPropertyById, which closes
its loading dialog with Get.back(); GetX's back() closes the current snackbar on
the way past and hit one still animating in, throwing LateInitializationError
out of SnackbarController. The dialog was then never dismissed and its barrier
sat over the whole screen swallowing every tap. It uses ScaffoldMessenger, which
GetX does not touch.

DealsService had no timeouts — the same omission MapService carried behind the
search. Accepting a counter awaits that call behind a spinner, so a stall left
the button spinning with no way to stop.

+++ aajoo_app_2026/lib/ui/screens_renter/property_details/open_property.dart
@@ -20,7 +20,7 @@ Future<void> openPropertyById(
   String? dealCode,
   String? dealFrom,
   String? dealTo,
-  int? dealPercent,
+  double? dealPercent,
   String errorTitle = 'Property',
 }) async {
   if (propertyId <= 0) return;
